[metasrv] refactor: move raft_txid to metasrv/raft/types/
